Skip hardware delays when running in Simics
There are sometimes delays in HWPs due to hardware delays that
need to be accounted for.  When we're running in Simics there is
no reason to wait because the hardware works instantly (in non-
modeled time).  Removing these extra delays can have a huge
benefit on the total runtime because there is a multiplicative
penalty between real time and simulated time.

Change-Id: I4f581ad2c028f250ed3fea00cc924e028f3d22d3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84522
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>