Fix ThreadPool Memory Corruption
One of the functions that ends up being called by threads spun by
ThreadPool has a potential memory corruption where a vector of data
structures can be searched while it's being populated by two different
threads. This commit changes the positions of mutex lock and ulock
calls in the vicinity to protect the population and search of the vector.

Change-Id: Ic1a5eb8928c1133ee2a99e2d9c04607be3c41b15
CQ: SW476649
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84561
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>