Update dimmspd used by explorers in axone simics
We want the explorer's SPD to contain valid i2c addresses for the PMIC
for simics. Currently simics has these as 0x90 and 0xC0. Normally we
ould just update simics directly and then update the simics relase but
there has been issues in the last few weeks getting onto a newer simics
release and we need these changes to test functionality that is being
added to aid with hardware bringup.

Change-Id: I1e29205cef59da986092bc8b609ebe27a89c5611
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84682
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>