Multicast - plat_target_handle_t Class
This commit transforms plat_target_handle_t into a class containing
TARGETING::Target (hostboot target) and multicast parameters. The
class supports both-way conversion to/from fapi2::Target via
constuctors and conversion operators.

The purpose of introducing additional parameters to the
plat_target_handle_t is to lay the groundwork for multicast
support in the nearest future. The additional parameters will allow
the multicast information to be passed to/from the fapi2 layer and
used in the SCOM device driver code to correctly assemble the multicast
SCOM address.

Change-Id: I601023d73ed936ae9db4a8ebbc793b5a1830832b
RTC: 203303
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85027
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>