Multicast - Actual Multicast Support
Support for muticast write. The SCOM address for multiast operations
needs to be morphed to include the multicast parameters (the multicast
bit (bit 1), the multicast op/type (bits 2-4) and multicast group (bits
5-7)); after the conversion, the hardware is responsible for decoding
the address and performing the series of SCOM operations. This commit
adds logic for adding multicast info into the SCOM address, the support
for multicast write, and some unit tests to test the address conversion
and the write operation.

Change-Id: Ia48ac1289bf07151d01f4add2b1cbf98b7b98049
RTC:203303
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85551
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>