Do not crash if we have no functional nodes
We hit a case where i2c errors caused the node target to get
marked as non-present.  This led to an assert crashing Hostboot
because it assumed we would always have a functional node0. The
change here is to allow this case to flow through so that we end
up with more recognizable error handling for the real failures
during boot.

Change-Id: Iaaaa4e531b677bda442da892cfbd6094e0897beb
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85169
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>