Dynamically generate ocmb cmd/rsp seq id
The current procedure code is using the sequence id portion
of the FW message to indicate the type of the command rather
than the command number. This makes error handling very
difficult because we can't detect which ocmb we are accessing
when we detect a failure. Create a function backed ekb attribute
to hold the sequence id so that every time we retrieve this attribute,
we increment the valuwe on the next get.

Change-Id: I93e0e7062d01ddf7ba6a69f89411a5b6bcf01471
RTC: 210371
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78767
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>