P10: istep 15.1
Although the P10 PM HWP's are not all working yet, the goal
of this story is to get everthing compiling in step 15 so that
the PM team can use our code to get the procedures working.

In order to allow hostboot work to procede, all PM HWP's will
be ifdef'd out by default.

RTC: 208832
Change-Id: Ife4f6053d20ffdfa9b5239487f7957645990a079
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85504
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>