Don't consider processor's OMI freq if it is 0 when freq restrictions
We have been told that processors must run at the same OMI frequency.
During bringup we saw that if the first boot was done with one of the
processors guarded out, and that processor was later unguarded. We would
fail the check that ensures these frequencies are the same, because
the proc that was guarded/unguarded would still be 0 where the other
would be set. This change will allow us to ignore processors with
unset frequency so we can boot far enough to set it. Later on the rule
will be enforced as well on subsequent boots.

Change-Id: I23950371f6f65750aacb07b28317ba11a3a4e8ab
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85832
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>