Multicast: Core Select Tests
This commit introduces unit tests for multicast Core Select
functionality and addresses some minor issues from previous commits.

Core Select allows to specify which cores in an EQ to send the SCOM
operation to via bits 16-19 in the multicast address. The requested
operation will then be performed on the selected cores for the selected
multicast group.

Change-Id: I63803e57178a6cc97ca4b2104ff0d00d12b7b7bf
RTC: 213303
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86178
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>