Make p10_update_ec_state match header prototype
The definition of p10_update_ec_state does not match the prototype
found in the header.  An additional parameter was added to the
definition (which is not used and is not seen in the ipl flow
document).  This was causing linker errors in hostboot environement.
This change removes the additional parameter.

Change-Id: I72ced8e1c061e763e394a0cb9c518e3399b89fe5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86192
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86245