P10 MRW parsing MEM Targets:  script updates: processMrw & Targets.pm
This commit deals with refactoring and modularizing the processMrw.pl
and Targets.pm scripts. This commit is not about the artifact that
gets produced by these files but the act of parsing an MRW file.

This commit is for targets MC, MI, MCC, DIMM, OCMB, MEM_PORT, OMI,
OMIC and PMIC. This is all the targets for story 208783.

Change-Id: I6e368188acd6dcf2ba345a9e750cb96edb4e35f8
RTC:208783
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86373
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>