Fix buffer overread in p10_pm_get_poundv_bucket
Prior to this commit, the function p10_pm_get_poundv_bucket had a
buffer overread bug when copying a ATTR_POUNDV_BUCKET_DATA_type into a
voltageBucketData_t variable; the size of the latter was used as the
length of the memcpy, but the former was smaller than that and the
copy would go out bounds on the read.

This commit temporarily fixes the problem by only copying as much data
as is allocated for the ATTR_POUNDV_BUCKET_DATA variable. A long-term
fix involving updating the layout of either or both structures should
be committed later.

Change-Id: I919e1830eca8c77d76d310e523f8fff376e073c9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86375
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Gregory S Still <stillgs@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86377
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>