Fix ipmifruinv calls for dimms in swift systems
We had a bug in the VPD map for the the DDR4_DDIMMs where we were
missing fields that were expected in the data collection for the
fruinventory. This commit adds the field we were missing and also
adds the PART_NUMBER field which the BMC is currently just leaving
blank.

Change-Id: I65dbfb0b79bef9d3a9a3bf2d6aa135c5df5ece56
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86420
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>