Adjust deadman timeout request time to be 10.5 seconds
We see intermittent deadman timeouts occuring in our testing. Upon
further inspection it was determined that value we were passing
to the SBE was being cast to a uint32_t and multiplied by up to
a factor of 400000 , which was resulting in an overflow of the
uint32_t. This was giving us a timeout lower than desired. This
commit lowers the requested timeout to be 10.5 seconds which will
not overflow the uint32_t.

Change-Id: I747db46a5d4a99d273b5604971d3db51f077be50
CQ: SW477557
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86513
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86566