Fixes to PMIC presence detection
While performing presence detection on the PMIC targets we need
to first read the parent OCMB's SPD to see what device address
the PMIC is on. There was a bug where we were attempting to read
the parent OCMB's spd without first checking if the OCMB is
present itself. This commit adds a check to ensure we dont attempt
i2c reads on devices that are not present. Also this commit adds
a check to make sure we do not attempt presence detection on GEMINI
ocmbs

Change-Id: I999189b3b97210bb37b7ba1fdb2d86658d770e36
CQ: SW480414
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86564
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>