Avoid commiting errors inside fapi2_i2c device driver
There are instances where we expect NACK errors returned from i2c
requests, for example when waiting for the ocmbs to declare themselves
"ready". In these cases we do not want to flood the error log
with expected I2C nack errors. To avoid this, instead of committing
retriable errors inside of the fapi2_i2c device driver when the MAX
RETRY ATTEMPTS is reached we will instead just add traces to the err
that gets returned to the caller that explain previous retryable errs
occured. In the case where we hit retryable errors and eventually hit
and non-retryable error the retryable error is committed in the driver
and the non-retryable error is returned to the caller. Also this commit
adds some timings to the error log so we can tell how long our attempts
took.

Change-Id: I9a5ee0bfe088444aa7fecd974f61514c40b320a6
CQ: SW480155
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86419
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>