Add support for NULL char in vendor log
SMART is adding a NULL char to indicate the last entry
added to the log.  This is useful when the log wraps.
A single NULL char indicates the log has wrapped.  More
than one NULL char in a row indicates the log has not wrapped.
Display the log in chronological order.

Change-Id: Ic0404b60ad617b7c0c6a09b710e54cb1736bd9d4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86630
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: TSUNG K YEUNG <tyeung@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>