backplane vpd
This commit adds the bp vpd file and reads that instead of
PVPD device read to populate the bp vpd and form slca for
bp proc and dimms. The slca index and rid for BP and SV is
yet to be created. Hdat goes till starting of MS vpd
creation with this commit

Change-Id: I09ba4a5c456e62fe4f2132d60bbf8855e7b02844
Signed-off-by: Sampa Misra <sampmisr@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87168
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>