Replace descriptions with JEDEC register names
For some NVDIMM errors, Hostboot generates a PEL record, that
contains a dump of NVDIMM controller registers.
The values of these registers can be decoded and analyzed using the
JEDEC specification JESD245, but it is difficult to do without standard
register names.
This changes replaces human readable descriptions with JEDEC specific
names in the errl plugin implementation.

Signed-off-by: Artem Senichev <a.senichev@yadro.com>