Replace descriptions with JEDEC register names
For some NVDIMM errors, Hostboot generates a PEL record, that
contains a dump of NVDIMM controller registers.
The values of these registers can be decoded and analyzed using the
JEDEC specification JESD245, but it is difficult to do without standard
register names.
This changes replaces human readable descriptions with JEDEC specific
names in the errl plugin implementation.

Signed-off-by: Artem Senichev <a.senichev@yadro.com>
Resolves #189
Change-Id: If7190772485530514e3ba21f48a04143de5a08d7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87583
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>