P10 MRW parsing misc targets: script updated: processMrw.pl & Targets.pm
Moved all of HB specific code out of Targets.pm into processMrw.pl. Mainly
moved the buildAffinity method, along with it's supporting routines, from
Targets.pm to the processMrw.pl file.

Renamed the buildAffinity to processTargets and the current processTargets
method was renamed to postProcessTargets because that what it seems to be doing.

Cleaned up the setCommonAttrForChiplet and added the setting of the CHIPLET_ID
which allowed the elimination of the setTargetChipletId method.  Also by
cleaning up the setCommonAttrForChiplet, the targets NMMU, NX, OCC, PEC and PHB
attributes fell out.

Reworked the way EEPROM and FAPI2 are done.  Added a way to verify if the
parent target has been processed before continuing on.  Removed all uses
of the ABUS, XBUS and OBUS.

Change-Id: Ibdf138d075f6998eb16725a8f24e30de9bb9866f
RTC: 215469
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87796
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>