Correctly Set ATTR_PG and Verify That HB and SBE Deconfig Match
Currently when HWAS State of an attribute is updated so is the ATTR_PG of that
attribute. This commit makes sure that ATTR_PG is only set with VPD and
Gard records.

Verify that the de-configuration logic calculated by SBE matches the result
of HB’s de-configurations.

Testcases to verify that new code catches when a target is deconfig by both
SBE and HB.

Change-Id: I5caebe9564455224fd054431cee73b96e3a4f6b7
RTC: 216231
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/86578
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>