EECACHE Presence Detect
Implement the functionality that drives presence detection based on
the contents of the EECACHE PNOR partition. During presence detection
for certain targets (EEPROM, OCMB, PMIC), we check the contents of the
header of the EECACHE partition and declare the target in question
functional or not based on whether the appropriate record is present
in the EECACHE header.

Change-Id: I6f2c34d5376c27dbb0971fbe79adea6028277fd4
RTC: 246952
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/88329
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>