Define Parent/Child Relationship for PAUC/OMIC targets (part 2)
There is a new relationship between PAUC targets and OMIC targets. The PAUC
can be the logical parent of an OMIC, and OMIC a logical child of a PAUC.

This commit updates the HWAS code for deconfiguring a PAUC parent and its OMIC
children. A missing PAUC parent will cause the OMIC children to deconfigure.
Missing OMIC children will not rollup the PAUC parent.

Change-Id: Ie74e914c0370b816b8a27f2afe5d98215532de45
RTC: 211222
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/88382
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>