ffdcExpander bash utility tool, written for December, 2019 Hackathon
The ffdcExpander tool will extract the various debug elements from a gzipped
SYSDUMP file, a SYSDUMP file or from a HB dump file with a given build tag.
The given dump files can be located in the current local directory or in a
directory elsewhere.  The artifacts that are produced are: Trace, Printk, Errl,
Errl with tool option 0X9000013E, Ps, Ps with backtrace, PageMgr and Bltrace.
These artifacts can be produced in the current local directory or redirected to
a directory elsewhere via option -d, default is the local directory.

Caller has the option of adding the artifacts attached to a defect via the -a
option.  If caller uses the -a option followed by a defect number, then the
caller will be prompted for their CQ user id and password.  Passwords will not
be displayed as they are typed.

An artifact that is produced is a file with the name of the FSP build tag.
Mainly as a FYI to the caller.  The contents of this file is the build tag,
FSP driver, FSP driver location and all the commands that ran to create the
other artifacts: Trace, Printk, etc.

Change-Id: Ib0dc202f4a7e77336ffb097b6647874dd7cebad0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89463
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>