Divorce global coverage request from individual artifact coverage decisions
Prior to this change, the HOSTBOOT_PROFILE var which indicates whether or not
to instrument coverage into the code could be unset by a blacklisted
module, which would then disable coverage from all of that module's subdirs.
This commit introduces a new flag, HOSTBOOT_PROFILE_ARTIFACT which is set
per individual module/artifact and removes the unwanted inheritance behavior.

Change-Id: I539a81053e59b9401c65b2437b3c3f9f9c7da1bf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/88789
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>