Stop unnecessarily regenerating build artifacts
This commit prevents the SBE image from being generated when
HBIBL has not been updated. This results in a ~60% performance
improvement in hb prime.

Also adds an environment flag called HB_SKIP_GEN_HCODE, which skips
regenerating the HCODE when it has already been created. This yields
a ~20% performance improvement in hb prime.

Change-Id: I9aae5d3a20d8522c381b624a555caf7571f5f89b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/88794
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>