Add current istep into TI SRC
Word4 of the SRC is defined to be the 'last progress code' for
FSP SRCs.  For Hostboot TIs, we currently leave that word zero.
This change will add the same data that we put into the scratch
register 5 into word4 so that we will have the failing istep for
any TI we encounter.

Change-Id: Iaf5ec835d45b1ea3a6ced20b5b7f7d07216c548e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/88833
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>