p10_ring_properties.H -- make perv_dpll_time RCLS_EKB_RINGS
The NEST and PAU DPLL configuration latches are wired to the TP time
ring for P10.  Normally time ring content is expected to come from
MVPD PDG, but that is not the case here.

Modify the perv_dpll_time ring properties entry to indicate that
content for this ring is expected to come from the initfiles
in EKB building through into the HW image.

Change-Id: Ia7991685bb9742d9884b2c693f3667021afd9027
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89118
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Claus M Olsen <cmolsen@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89121
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>