Handle automatic interleaving between OMI sub-channels
The memory channel is defined as a single MCC, and all of the
memory on that channel is part of a single address space.  That
means that both OCMBs on the same MCC share an address space.
While this isn't the same mechanism as true interleaving, it
looks the same to the code consuming HDAT, so we need to set
the sharing flags appropriately.

Change-Id: Ic270582ced92614cf21ff7be0ef6f97c93ab2514
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89340
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>