SMF: Distribute 0 Mem By Default
When the amount of SMF memory to be distributed is not specified
in skiboot/NVRAM, we don't turn SMF off, since the leg to distribute
0 memory is not executed. That causes issues in istep 15 where we
expect SMF BAR to be nonzero, and we assert since SMF mode is still
on.

This commit factors out the call to distribute SMF mem into general
flow and makes the default amount to distribute 0, which will force
SMF to be turned off.

Change-Id: I48c8d21bee67c4aafd0c21cabc449037def841b4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/90377
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>