Add base support for MCTP initialization on the Host
This commit adds in the code that allows Hostboot to
initialize the host side of the MCTP interface we are
setting up in order to do PLDM traffic. The libmctp.so
object will get loaded by initservice as a base service.
When the object is loaded it will spawn daemon threads that
will poll on the KCS status register and handle KCS cmds
that come up.

RTC: 215965
Change-Id: I7b94983b849bcff1cc0a30cbd911b79f8e0bb8da
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/90325
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>