Mvpd-Ekb ringId gap-discontinuity support
Making support for the ringId gap that needs to exist between the end
of the Mvpd RingID group and the beginning of the Ekb RingID group
in order to be able to append Mvpd rings to the end of the Mvpd
RingID group without affecting the already assigned RingID enum for
any of the other existing rings.

Note that the purpose of any RingID grouping, or order, is to:
- ease debugging the ring images and output traces of ipl_customize,
- ease the maintaining of the ringId header files.

Enforcing ringId groups comes with an increase in code complexity
to enable conversion between ringId and rpIndex (i.e., the index of
of the RING_PROPERTIES list).

Several bugs were found in implementing the above support as there has
been a built-in assumption in the PPE/Cronus customized putRingUtils
and ring_traverse codes that the ringId is a uint8 type but which we
are now exceeding the limit of with this group enforcement.

Also added updated mvpd.dat, based on scandef ver 910007, to better
match EKB engd's version (which at the time of this merge seems to be
consistent with 910007).
- Also added an README file with details about mvpd.dat

Key_Cronus_Test=XIP_REGRESS_SBE_QME

Change-Id: I59eddeff34c3dedf23b092b0cf89ef422e8bb753
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/91255
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/91400
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Nicholas E Bofferding <bofferdn@us.ibm.com>