Re-organize PLDM directories
We know that the PLDM code will need to be broken up into multiple
modules. At minimum, there will be a base, extended, and runtime
image. We must have base and extended because we do not want to fit
all of the response behavior we need to support in the Hostboot Base
Image (HBB) PNOR partition. This is similar behavior to the secureboot
modules so I based the structure off of that. All code from github is
now in pldm/extern. The base image will contain the pldmrp. The extended
behavior, which will be included in the Hostboot Extended Image (HBI
PNOR partition, will potentially contain all but certainly some of the
response functionality. Other directories under pldm will not contain
makefiles and instead will be referenced from either the base, extended,
or runtime makefiles.

Change-Id: I646ae19315b6d67815f7f7214f2c27e6093cb678
RTC: 208801
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/91705
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>