Change behavior of hb prime --release
The --release flag is used to distribute some aspect of
Hostboot to external consumer.  The primary usage is to
generate the files for a hostboot release into fips cmvc.
Previously it defaulted to distributing all of our targets,
including OpenPOWER specific artifacts that don't always
exist.  This change will make --release distribute only
a subset of possible targets : fsp, tools, errlparser.

Additionally, the fsp distribution target was purged of any
ode mk directives as they were only necessary to create the
ode-based simics environment that is no longer being used.

Also made a tweak to the image makefile to support a new SBE
image when it shows up.

Change-Id: Ibb8c53ee1ca0b0eef0ea703710bb823dc3d765c8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/92053
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>