Convert DECONFIG to DELAYED_DECONFIG on FSP systems
There is a problem if Hostboot sends an error log with a regular
DECONFIG option and then fails the istep.  In this case, HWSV has a race
in the order that it processes the deconfig request and the istep fail
message.  If the istep failure happens before the deconfig takes place
(inside of HWSV) it will be treated as a hard failure.  The correct
behavior is that when there is a deconfig during an istep failure that
we trigger a reconfig loop.  Therefore this issue results in an IPL
failure instead of a reconfig loop, which is not correct.  To avoid this
issue we will always use DELAYED_DECONFIG while inside our IPL context
which will force HWSV to act the way we want them to.

Change-Id: I09cb42e23293e3886f3f3492bec73b796c74fa0f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94135
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>